refactor(relocate-files): relocated the plugin_manager
